refactored utils into several smaller files
